+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Try to find locally-install phantomjs first.
 
+* `Session$new()` now checks that the phantomjs binary was built with ghostdriver support.
+
 
 # 1.0.3